% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcAgeSchedules.R
\name{calcAgeSchedules}
\alias{calcAgeSchedules}
\title{Calculate age-schedule information.}
\usage{
calcAgeSchedules(sID)
}
\value{
sID object with additional information.
}
\description{
Calculate age-schedule information.
}
\details{
Computes the length-at-age, weight-at-age
maturity-at-age,survivorship, selectivity, and SSB per recruit.
}

