% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcBoSteepness.R
\name{calcBoSteepness}
\alias{calcBoSteepness}
\title{calculates unfished spawning biomass and steepness given MSY and FMSY}
\usage{
calcBoSteepness(sID)
}
\value{
sID object
}
\description{
calculates unfished spawning biomass and steepness given MSY and FMSY
}

