% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcObjectiveFunc.R
\name{calcObjectiveFunc}
\alias{calcObjectiveFunc}
\title{Compute the objective function.}
\usage{
calcObjectiveFunc(sID)
}
\value{
sID object
}
\description{
Compute the objective function.
}
\details{
Computes the objective function for Importance Sampling or
			non-linear parameter estimation.
}

