% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY.R
\docType{package}
\name{catchMSY}
\alias{catchMSY}
\alias{catchMSY-package}
\title{catchMSY: A package for implementing the age-structured integrated catch-MSY method.}
\description{
Estimates management-oriented parameters (MSY and Fmsy), requires at least a time series of catch data and assumptions about biological parameters, and can integrate an abundance index, biomass survey, mean length time series, and length composition data, if available. Underlying age-structure allows for better accounting of age-structured processes. Uses the sampling-importance-resampling (SIR) algorithm for Bayesian analysis to sample the posterior distributions of parameters.
}
\author{
Steve Martell (martell.steve@gmail.com), Merrill Rudd (merrillrudd@gmail.com)
}
\keyword{age-structured,}
\keyword{catchMSY,}
\keyword{integrated,}
\keyword{management-oriented}
\keyword{parameters}

