% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_sID.R
\name{new_sID}
\alias{new_sID}
\title{Generate a new stock ID object}
\usage{
new_sID(id = "Stock Label", age = 1:10, linf = 100, winf = 5,
  vbk = 0.2, to = 0, a = 5e-06, b = 3, mat_age_50 = 2,
  mat_age_95 = 3, dfile = "")
}
\arguments{
\item{id}{Name of Stock}

\item{dfile}{Name of file containing time series data}
}
\value{
A template for a stock class object.
}
\description{
Generate a new stock ID object
}
\details{
This function returns a list object containing all the necessary
         information to run the catchMSY routines.
}
\examples{
#myStock <- new_sID()
#names(myStock)
}

