% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY_graphics.R
\name{plot.stockID}
\alias{plot.stockID}
\title{S3 class methods for providing summary plots}
\usage{
\method{plot}{stockID}(sID, .THEME = NULL)
}
\description{
S3 class methods for providing summary plots
}

