% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runAgeStructuredModel.R
\name{runAgeStructuredModel}
\alias{runAgeStructuredModel}
\title{Run Age Structured Model on sID object.}
\usage{
runAgeStructuredModel(sID)
}
\arguments{
\item{sID}{A stock object.}
}
\value{
A modified \code{sID} containing the output from an age-structured
        model.
}
\description{
Run Age Structured Model on sID object.
}
\details{
The age-structured model is conditioned on the historical catch data
(in weight).  The catch equation assumes that both
fishing mortality and natural mortality are occuring simultaneously, and the
instantaneous fishing mortality rate is found by solving the Baranov catch
equation \eqn{C =  F/Z*(1-exp(-Z))*B}, for \code{F}.
}

