% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSIR.R
\name{runSIR}
\alias{runSIR}
\title{Run Sample Imporance Resample routine.}
\usage{
runSIR(sID, n = 100, nc = 1)
}
\arguments{
\item{sID}{is the stock object}
}
\description{
Run Sample Imporance Resample routine.
}

