% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catchMSY_searchTools.R
\name{sir.sid}
\alias{sir.sid}
\title{Importance Sampling}
\usage{
sir.sid(sID, ncores = 1)
}
\arguments{
\item{sID}{Stock ID object}

\item{nc}{Number of cores for parrallel processing.}
}
\description{
Importance Sampling
}

