% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runModel.R
\name{solver}
\alias{solver}
\title{Envoke a non-linear search routine to estimate model parameters.}
\usage{
solver(sID)
}
\value{
sID object
}
\description{
Envoke a non-linear search routine to estimate model parameters.
}

