% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY.R
\docType{package}
\name{catchMSY}
\alias{catchMSY}
\alias{catchMSY-package}
\title{catchMSY: A package for implementing the catch-MSY method.}
\description{
\packageDESCRIPTION{catchMSY}
}
\section{catchMSY functions}{

The foo functions ...
}
\author{
\packageAuthor{catchMSY}
}
\keyword{catchMSY}

