% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catchMSY_getFt.R
\name{getFt}
\alias{getFt}
\title{Use Newtons method to find instantaneous F}
\usage{
getFt(ct, m, va, wa, na)
}
\value{
instantaneous fishing mortality rate F
}
\description{
Use Newtons method to find instantaneous F
}

