% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catchMSY_stockID.R
\name{new_sID}
\alias{new_sID}
\title{Generate a new stock ID object}
\usage{
new_sID(id = "Stock Label", age = 1:10, linf = 100, winf = 5,
  vbk = 0.2, to = 0, a = 5e-06, b = 3, ah = 2, gh = 0.2, m = 0.2,
  fmsy = 0.15, msy = 1, dfile = "")
}
\arguments{
\item{id}{Name of Stock}

\item{age}{A vector of integer ages where the terminal age is a plus group.}

\item{linf}{Asymptotic length for von Bertalanffy growth}

\item{winf}{Asymptotic weight.}

\item{vbk}{von Bertalanffy growth coefficient.}

\item{to}{theoretical age at 0 length (years).}

\item{a}{allometric scaler in length-weight relationship \eqn{w = a L ^b}}

\item{b}{allometric power parameter in length-weight relationship}

\item{ah}{age at 50\% maturity}

\item{gh}{standard deviation in age at 50\% maturity}

\item{dfile}{Name of file containing time series data}
}
\value{
A template for a stock class object.
}
\description{
Generates a template sID object for new assessments.
}
\details{
This function returns a list object containing all the necessary
         information to run the catchMSY routines.
}
\examples{
myStock <- new_sID()
names(myStock)
}

