% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY_searchTools.R
\name{runSIR}
\alias{runSIR}
\title{Run Importance Sampling}
\usage{
runSIR(sID, n = 100, nc = 1)
}
\arguments{
\item{sID}{is the stock object}

\item{n}{number of samples}

\item{nc}{number of cores}
}
\description{
Run Importance Sampling
}

