% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catchMSY_searchTools.R
\name{sample.sid}
\alias{sample.sid}
\title{Generate prior samples for importance sampling.}
\usage{
sample.sid(sID, n = 100)
}
\description{
Draw random samples from specified parameter ranges in the
stock ID object.
}

