% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY_searchTools.R
\name{sir.sid}
\alias{sir.sid}
\title{Importance Sampling}
\usage{
sir.sid(sID, selex = FALSE, ncores = 1)
}
\arguments{
\item{sID}{Stock ID object}

\item{selex}{Boolean flag to turn off search across sel50 parameter. 
Set to TRUE when including the prior on sel50.}

\item{nc}{Number of cores for parrallel processing.}
}
\description{
Importance Sampling
}

