% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY_searchTools.R
\name{sir.sid}
\alias{sir.sid}
\title{Importance Sampling}
\usage{
sir.sid(sID, ncores = 1, selex = FALSE)
}
\arguments{
\item{sID}{Stock ID object
Set to TRUE when including the prior on sel1 (sel50).}

\item{selex}{Boolean flag to turn off search across sel1 (sel50) parameter.}

\item{nc}{Number of cores for parrallel processing.}
}
\description{
Importance Sampling
}

