% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchMSY_nls.R
\name{solver}
\alias{solver}
\title{Envoke a non-linear search routine to estimate model parameters.}
\usage{
solver(sID, selex = FALSE)
}
\value{
sID object
}
\description{
Nonlinear search routine for estimating model parameters given data.
}
\details{
The solver function relies on 'optim' to conduct a non-linear search
for estimating model parameters (M, FMSY, MSY).
}

